refactor(frontend): complete features/home/ui module (FR-006 US2.2)
