+++ frontend/src/app/(marketing)/page.tsx
@@ -1,4 +1,4 @@
-import { HomeClient } from '@/components/home/HomeClient';
+import { HomeClient } from '@/features/home/ui/HomeClient';
 
 export default function Home() {
   return <HomeClient />;
