+++ frontend/src/features/home/ui/BottomNav.tsx
